add changes to init_db.sql file and to exception.

+++ src/main/resources/init_db.sql
@@ -1,7 +1,7 @@
 CREATE DATABASE `manufacturer_db`;
 
 CREATE TABLE `manufacturers` (
-  `id` int NOT NULL AUTO_INCREMENT,
+  `id` bigint NOT NULL AUTO_INCREMENT,
   `name` varchar(45) NOT NULL,
   `country` varchar(45) NOT NULL,
   `is_deleted` tinyint NOT NULL DEFAULT '0',
@@ -9,7 +9,7 @@ CREATE TABLE `manufacturers` (
 ) ENGINE=InnoDB AUTO_INCREMENT=5 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci
 
 CREATE TABLE `drivers` (
-  `id` int NOT NULL AUTO_INCREMENT,
+  `id` bigint NOT NULL AUTO_INCREMENT,
   `name` varchar(45) NOT NULL,
   `license_number` varchar(45) NOT NULL,
   `is_deleted` tinyint NOT NULL DEFAULT '0',
